+++ CHANGELOG.md
@@ -7,9 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
-  - ?
+  - `ktlintRuleset` configuration to provide 3rd party ktlint rules (#71)
 ### Changed
   - Update Kotlin to `1.3.30` version
+  - Deprecated providing 3rd party ktlint rules via extension (#71)
 ### Fixed
   - Proper lazy adding ktlint dependency (#219)
 ### Removed